IGN-7871 WIP PKCS#11 keystore support, cont'd
We're hitting an interesting blocker with the YubiKey 5's signing cert
slot. This slot requires a PIN (aka cert password) on signing
operations, it's not good enough to just unlock the keystore at the
outset via its PIN (aka keystore password). The current logic throws
`java.security.ProviderException:
sun.security.pkcs11.wrapper.PKCS11Exception: CKR_USER_NOT_LOGGED_IN`
down in `module-signer`'s `ModuleSigner.asymmetricSignature`. Working
backward, it's clear it successfully logged into the keystore and
fetched the private key "stub," but there is a PIN challenge again
arising during `Signature.sign` AFAICT.

`SignModule` does not have any sort of `PasswordCallback` handler
specified in `getKeyStore`. In theory that *could* feed the YK5 a
password on every signing operation. However, it's not clear that this
would be necessary or appopriate for every HSM. And if it was, would
there need to be a distinct implementation for every flavor of HSM?

Moreover we're conflating the keystore PIN with the signing cert PIN.
`certPassword` is doing double duty. That should be fixed, albeit
that's a relatively easy fix.

Ref: <https://docs.oracle.com/en/java/javase/17/security/pkcs11-reference-guide1.html>