IGN-9137 `ignitionModule` config change
`generator-core` should accommodate module dependencies with the new
required/not required semantics for Ignition 8.3, while remaining
backward-compatible for Ignition 8.1 modules.

The new `optionalModuleDependencies` are consumed by the
`writeModuleXml` task, writing a `depends` tag as it does for
the legacy `moduleDependencies`. The difference is that it adds the
tag attribute `required="false"`.

Conversely `requiredModuleDependencies` yield a `depends` tag with
attribute `required="true"`.

Note the semantics of the old `moduleDependencies` aren't exactly
the same as `optionalModuleDependencies` per my reading of RFC-54.
We don't want the generator to implicitly treat (8.1ish)
`moduleDependencies` like the latter, writing a `required="false"`
attribute and possibly choking an 8.1 gateway that is not expecting
it.